DNM:Manually retrigger 0.10.3 gpu wheel deploy jobs
For the 0.10.3 release the qiskit-aer-gpu release jobs failed, this
was caused by the addition of python 3.10 which uses the manylinux2014
image (for compat with numpy and scipy) to build wheels. The
manylinux2014 image uses centos7 as it's base instead of centos6. The
job failed when trying to build the python 3.10 wheel because we
installed the cuda centos6 package on centos 7 and it wasn't compatible.
This commit fixes this by splitting out the 3.10 job and updating the cuda
package to be compatible with centos7.

This commit should not be merged to a real branch, it's setup to
manually trigger the release jobs and publish wheels to pypi inplace of
the failed job. A real PR integrating this fix will be opened after the
wheels are published